backend/scraper/requirements.txt
crawlee[playwright]==0.3.0
python-dotenv==1.0.1
redis==5.0.4
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
lxml==5.2.2
